backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
kafka-python==2.0.2
elasticsearch==8.12.1
httpx==0.27.0
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
